+++ include/pangolin/triangle_counter/cudamemcpy_tc.hpp
@@ -21,5 +21,5 @@ public:
 	virtual void read_data(const std::string &path) override;
 	virtual void setup_data() override;
 	virtual size_t count() override;
-	virtual size_t num_edges() override { return hostDAG_.num_edges(); }
+	virtual uint64_t num_edges() override { return hostDAG_.num_edges(); }
 };
